Remove legacy `max-skip-slots` checks (#4403)
Remove `max-skip-slots` checks when processing blocks.
This was legacy code which was previously used in the Medalla testnet to sync to the correct fork.
With the addition of checkpoint sync which allows us to sync to any arbitrary fork, this is no longer a necessary feature, so it has been removed for simplicity.

The CLI flag and checks for attestation processing have been retained as it still may have uses in DoS protection.