added debounce to log (#4269)
[#4259](https://github.com/sigp/lighthouse/issues/4259)

debounce spammy `Unable to send message to the beacon processor` log messages

We could potentially debounce other logs that have the potential to be "spammy".

After some feedback we decided to additionally add the following change:

create a newtype wrapper around `mpsc::Sender<BeaconWorkEvent<T>>`. When there is an error on the try_send method on the wrapper, we increase a counter metric with one label per work type.